Make version number work well with build tools
In 3d15088, a `__version__` variable was introduced with a conditional
suffix, e.g. ` (in development at /path/to/projectroot)'`.

Turns out that `__version__` may be consumed by build tools, too.
Inserting such a suffix may break the build in the best case (because it
detects the PEP 440 violation, like pybuild does) or it may silently
leak the suffix into production in the worst case.

Move the suffix out of `__version__` into the `--version` handler, which
now checks the presence of `pyproject.toml` separately and then appends
the suffix if need be.